Terminando las validaciones y comprobaciones del MVC de las personas

+++ controller/person_controller.py
@@ -23,18 +23,25 @@ class ControladorPersona:
                 if personas:
                     self.view.mostrar_mensaje("\n--- Lista de Personas ---")
                     self.view.mostrar_personas_id(personas)
+                    
                     while True:
                         self.view.mostrar_mensaje("\n--- Actualizar Persona ---")
                         id_persona = self.view.obtener_id_persona("actualizar")
+                        
                         if self.modelo.validar_id_persona(id_persona):
-                            nuevos_datos = self.view.obtener_datos_persona()
-                            self.modelo.actualizar_persona(id_persona, nuevos_datos)
+                            campo = self.view.seleccionar_campo_actualizar()
+                            while True:
+                                nuevo_valor = self.view.obtener_nuevo_valor(campo)
+                                if nuevo_valor:
+                                    self.modelo.actualizar_campo(id_persona, campo, nuevo_valor)
+                                    break
                             break
                         else:
                             self.view.mostrar_mensaje("ID no válido. Intente de nuevo.")
                 else:
                     self.view.mostrar_mensaje("No hay personas registradas.")
-            elif opcion == "4":
+
+            elif opcion == "4":  # Eliminar persona
                 personas = self.modelo.ver_personas()
                 if personas:
                     self.view.mostrar_mensaje("\n--- Lista de Personas ---")
@@ -43,17 +50,20 @@ class ControladorPersona:
                         self.view.mostrar_mensaje("\n--- Eliminar Persona ---")
                         id_persona = self.view.obtener_id_persona("eliminar")
                         if self.modelo.validar_id_persona(id_persona):
-                            self.modelo.eliminar_persona(id_persona)
+                            confirmacion = self.view.confirmacion_eliminacion()
+                            if confirmacion.strip().upper() == "SI":
+                                self.modelo.eliminar_persona(id_persona)
+                            else:
+                                self.view.mostrar_mensaje("Eliminación cancelada.")
                             break
                         else:
                             self.view.mostrar_mensaje("ID no válido. Intente de nuevo.")
                 else:
                     self.view.mostrar_mensaje("No hay personas registradas.")
+
                 
             elif opcion == "5":
                 self.view.mostrar_mensaje("Saliendo del sistema...")
                 break
-            else:
-                self.view.mostrar_mensaje("Opción no válida. Intente de nuevo.")
